requirements.txt
scrapy-splash
psycopg2
scrapy-user-agents

